renamed npc events file to avoid duplicate names
