+++ events/npc_event.js
