+++ index.js
@@ -5,7 +5,7 @@ import { initialize_items, items_list } from './initializers/items.js';
 import { initialize_djinni, djinni_list } from './initializers/djinni.js';
 import { initialize_enemies, enemies_list } from './initializers/enemies.js';
 import { initialize_maps, load_maps, maps } from './initializers/maps.js';
-import { set_npc_event, trigger_npc_dialog } from './events/npc.js';
+import { set_npc_event, trigger_npc_dialog } from './events/npc_event.js';
 import { initialize_menu } from './screens/menu.js';
 import { TileEvent } from './base/tile_events/TileEvent.js';
 import { Debug } from './debug.js';
